simplified and prepared SOAP protocol
