+++ package.json
@@ -20,7 +20,9 @@
   "engines": {
     "node": ">=0.4.7"
   },
-  "dependencies": {},
+  "dependencies": {
+    "soap": "^0.9.1"
+  },
   "devDependencies": {
     "mocha": "~2.2.4"
   },
