+++ test/input/chinese-large-html.txt
@@ -196,76 +196,5 @@
     </span><p> 他们又不小心把这种新东东用到了卫生巾上 </p><span>
     </span><p> 就创造出了 液体卫生巾 </p><span>
     </span><p> 卫生巾的历史 </p><span>
-    </span><p> 就这么一不小心又被改写了 </p><span>
-    </span><p> ▼ </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615160800_138.gif" alt="" width="400" height="400" title=""> </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615160823_616.jpg" alt="" width="400" height="400" title=""> </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615160825_999.jpg" alt="" width="400" height="225" title=""> </p><span>
-    </span><p> 要问用上这款液体卫生巾是什么感觉？ </p><span>
-    </span><p> 就像那些穿越客一样： </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615160856_645.gif" alt="" width="400" height="400" title=""> </p><span>
-    </span><p> 具体是什么感觉呢？ </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615160918_56.gif" alt="" width="400" height="475" title=""> </p><span>
-    </span><p> 不管姨妈什么时候来： </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615160934_360.gif" alt="" width="400" height="475" title=""> </p><span>
-    </span><p> 不管你是女王，还是打算穿越去当女王： </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615160951_110.gif" alt="" width="400" height="475" title=""> </p><span>
-    </span><p> 不管你是女神，还是打算穿越去当女神： </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615161007_223.gif" alt="" width="400" height="475" title=""> </p><span>
-    </span><p> 不管你是女汉子，还是穿越去当女汉子： </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615161025_653.gif" alt="" width="400" height="475" title=""> </p><span>
-    </span><p> 不管你穿越，还是不穿越： </p><span>
-    </span><p>
-        <img src="http://image.meihua.info:808/editor/2015/06/15/20150615161047_762.jpg" alt="" width="400" height="400" title=""> </p><span>
-    </span><p> 部分资料来自果壳网文章《卫生巾是怎么来的》 </p><span>
-    </span><p> （作者roseVsthorn） </p><span>
-    </span><p>
-        <strong>思路 </strong>
-    </p><span>
-    </span><p> 很多人一听到液体卫生巾，第一反应都是“哇，好像很NB，不过这货是什么？”，甚至有人以为它是液态的用来抹的。 </p><span>
-    </span><p> 而我的任务，就是要说清楚这货到底是什么，比传统卫生巾NB在哪里，并且设法让人印象深刻。毕竟事关生理卫生，不把这样一个全新的产品概念阐述清楚，很多人是不会贸然尝试的。而这事儿，并非靠几个高大上的KV、几句貌似走心的漂亮话可以解决。至少我没这本事。 </p><span>
-    </span><p> 所以我选择了一个已被多年证实有效的方法，来跟消费者做深度沟通，那就是： </p><span>
-    </span><p>
-        <strong>通过对比，来建立产品的NB感。</strong>
-    </p><span>
-    </span><p>
-        <strong>具体到液体卫生巾上，有两条路可走。</strong>
-    </p><span>
-    </span><p> 一条路是横向的，也就是和市面上同类产品做对比——这也是宝洁家惯用的广告模式，从洗发水、面霜、纸尿裤到卫生巾，都是这么干的。护舒宝早就为此做了一些很直观的视频和动图。选这条路的话，走得越远，恐怕是越白费力气。是可以从表现形式上入手，想一些比较花俏、新奇、看起来有idea的对比实验方式，但这没有意义。我认为，还是直给更有力。 </p><span>
-    </span><p> 另一条路是纵向的，也就是去讲女性卫生用品的演变史，最后有理有据地推导出液体卫生巾就是迄今最能吸收最贴身最透气最无感最舒爽的。这事儿护舒宝还没做。路挺宽的看上去，我想要试着走一走。 </p><span>
-    </span><p> 于是，在做了大量功课及体验之后，就有了这样一个创意概念： </p><span>
-    </span><p>
-        <strong>妹纸穿越必备（因为液体卫生巾是有史以来最好的卫生巾）</strong>
-    </p><span>
-    </span><p>
-        <strong>不穿越更要常备（因为液体卫生巾是有史以来最好的卫生巾）</strong>
-    </p><span>
-    </span><p> 看起来很扯，但我是很严肃地在扯，而且逻辑严密，不服来辩。 </p><span>
-    </span><p> 然后就有了这样一篇《生理期穿越指南》。 </p><span>
-    </span><p> 其实，卫生巾的演变史，就是人类进步史的一个小小的、有力的注脚。有的人爱怀念以前慢什么的。拉倒吧，你知道从前慢，可你知道从前烂吗？哪曾有过什么黄金时代，现在就是有史以来最好的时代。不信你用液体卫生巾感受一下。 </p><span>
-    </span><p>
-        <strong>作者|陈野</strong>
-    </p><span>
-    </span><p>
-        <strong>本文由NewbeeRen(ID:iamnewbeeren) 授权梅花网转载，版权归原作者所有。</strong>
-    </p><span>
-    </span><p>
-        <strong>如需转载请注明作者、出处，并保留文章在梅花网（www.meihua.info）的完整链接，谢谢！</strong>
-    </p><span>
-    </span><p>
-        <span>
-            <strong>
-                <img src="http://image.meihua.info:808/editor/2015/06/15/20150615161229_504.jpg" alt="">
-                <br></strong>
-        </span>
-    </p><span>
+    </span><p> 就这么一不小心又被改写了 </p>
 </span></div>
